Simplified end to end test
